introduction: add an explanation of "hello world"
The introduction of the error assertion to the hello world program is
probably worth explaining at least somewhat, because throughout the rest
of the tutorial it will be used on some calls and not on others, and I
would rather not have anyone getting confused as to why.

To address this, I added a section which breaks down the hello world
program in detail, and includes a brief explanation of error handling.