introduction: Rewrite switch & match to new syntax
Basically just a mechanical change updating the examples to use the new
syntax with "case" for switch and match expressions.

Tested by copy-pasting each sample that contains switch and/or match
into a test.ha file, and running "hare run test.ha" (obviously the
driver is built from the switchcase branch, not master).

An outstanding todo item is to re-consider the order of the sections to
introduce compound expressions and yield they're used in switch/match
statements.

Also fixes a bug in the "Switch statements" section: The second use of
switch should have been "switch (color)", not "switch (x)".

Signed-off-by: Thomas Bracht Laumann Jespersen <t@laumann.xyz>