introduction: Introduce compound expressions and yield earlier
The concept of compound expressions and the use of "yield" is now
introduced right after "if" in the section on control flow constructs.
This is so we can expect the reader to have seen how to use yield before
"switch" and "match" are introduced. The "if" section is reduced
slightly to avoid the use of yield (but still talks about "if" being an
expression).

The sections on "switch" and "match" are also updated to talk about the
syntax correctly.

Signed-off-by: Thomas Bracht Laumann Jespersen <t@laumann.xyz>