Proofreading
The goldmark typographer converts `--` into an en dash rather than an em
dash, which it uses `---` for. To fix this, I replaced all occurrences
of `--` with `&mdash;` as it was already used throughout the site.

I also fixed a typo (unifiing -> unifying), some noun-verb agreement
issues (provides -> provide, reduces -> reduce), and an instance of
"it's" being used instead of "its" for the possessive.

(Apologies for any mistakes, this is my first time sending in a
patch via email! Other than the tutorial, of course, which was very helpful)